normalize issuer before testing during validation
Section 2.1.2 for the OpenID Connect Discovery 1.0 spec
indicates that identifiers without a scheme component should
be assumed to use the https scheme and be normalized as such.
In order to support ODIC providers which rely on this
normalization, check the issuer string during Provider setup
and add the https scheme to the provider issuer, if necessary,
during setup and validation.